1. Fix the build error caused by the modification for Compiler 6 2. Revise startup_M0518.s and retarget.c file to fix Semihost build issue 3. Modify settings of Keil project to spped up compilation 4. Revise copyright format for gen. chm file 5. Fix download issue of all IAR Projects.

+++ SampleCode/StdDriver/I2C_Wakeup_Slave/main.c
@@ -7,8 +7,9 @@
  *           Show how to wake up MCU from Power-down mode through I2C interface. 
  *           This sample code needs to work with I2C_Wakeup_Master.
  * @note
- * Copyright (C) 2014 Nuvoton Technology Corp. All rights reserved.
+ * @copyright SPDX-License-Identifier: Apache-2.0
  *
+ * @copyright Copyright (C) 2014 Nuvoton Technology Corp. All rights reserved.
  ******************************************************************************/
 #include <stdio.h>
 #include "M0518.h"
@@ -71,7 +72,7 @@ void PWRWU_IRQHandler(void)
         /* Clear system power down wake-up interrupt flag */
         CLK->PWRCON |= CLK_PWRCON_PD_WU_STS_Msk;
 
-        printf("Power-down Wake-up INT 0x%x\n", ((CLK->PWRCON) & CLK_PWRCON_PD_WU_STS_Msk));
+        printf("Power-down Wake-up INT 0x%x\n", (unsigned int)((CLK->PWRCON) & CLK_PWRCON_PD_WU_STS_Msk));
     }
 }
